Rename Footer companent
